Update 1.1.2 Patch 2
- Добавлен новый HUD (HP, XP, GOLD)
- Добавлен выбор класса
- Добавлены настройки и сайт с донатом и новостями
- Прочие изменения

+++ Assets/ProjectUse/Scripts/Interfaces/IHealth.cs
@@ -2,7 +2,7 @@ using System;
 
 public interface IHealth
 {
-    public int Health { get; }
+    public float Health { get; }
 
     public EventHandler<int> TakeDamage { get; }
     public EventHandler<int> TakeHeal { get; }
